Move env var converter to its own file
